requirements.txt
openai 
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
faiss-cpu
pandas
numpy
python-dotenv
fastapi[all]
mysql-connector-python